requirements.txt
vllm>=0.6.0
torch>=2.0.0
pillow>=10.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
